+++ docs/zh/guide/other/changelogs/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.13.10](https://github.com/fast-crud/fast-crud/compare/v1.13.9...v1.13.10) (2023-05-31)
+
+### Bug Fixes
+
+* 修复search-middle插槽错位问题 ([6423c14](https://github.com/fast-crud/fast-crud/commit/6423c14b037c18f1aea87b975b8f323db2151fb3))
+
+## [1.13.9](https://github.com/fast-crud/fast-crud/compare/v1.13.8...v1.13.9) (2023-05-31)
+
+### Bug Fixes
+
+* 恢复search插槽 ([12580ef](https://github.com/fast-crud/fast-crud/commit/12580ef18d30b69c6a995d6c16b03bfc44c411f1))
+* 修复antdv文件上传限制数量的bug ([8b14ba3](https://github.com/fast-crud/fast-crud/commit/8b14ba3a45f90a11222cc751b2ca173e212bc666))
+
+## [1.13.8](https://github.com/fast-crud/fast-crud/compare/v1.13.7...v1.13.8) (2023-05-22)
+
+### Performance Improvements
+
+* **form:** 支持conditionalRender ([270fd52](https://github.com/fast-crud/fast-crud/commit/270fd52e7ffeab605aa89f44dc1b1e8f132ce393))
+
+## [1.13.7](https://github.com/fast-crud/fast-crud/compare/v1.13.6...v1.13.7) (2023-05-19)
+
+### Bug Fixes
+
+* 修复rowhandle 排列不整齐的问题 ([ff644b1](https://github.com/fast-crud/fast-crud/commit/ff644b11b91c58295692fc0874dc4a3d743eb6df))
+* 修复useFormWrapper 报inject错误的bug ([fbaf8b8](https://github.com/fast-crud/fast-crud/commit/fbaf8b8e3f745cb00df40e270928bd8152a61e66))
+
 ## [1.13.6](https://github.com/fast-crud/fast-crud/compare/v1.13.5...v1.13.6) (2023-05-13)
 
 **Note:** Version bump only for package fast-crud-root